[btree] Factor out benchmarks
