+++ src/tests/btree.rs
@@ -17,7 +17,6 @@ use nom::{number::complete::*, IResult};
 use rand::prelude::*;
 use rand::SeedableRng;
 use std::collections::BTreeSet;
-use std::fs::File;
 use std::io;
 use std::io::{Cursor, Read, Write};
 use std::marker::PhantomData;
@@ -100,7 +99,7 @@ impl<V: Unpack> NodeVisitor<V> for ResidencyVisitor {
     }
 }
 
-fn get_stats<V: Unpack>(bm: &Arc<BlockManager>, root: u64) -> Result<TreeStats> {
+fn get_tree_stats<V: Unpack>(bm: &Arc<BlockManager>, root: u64) -> Result<TreeStats> {
     let walker = BTreeWalker::new(bm.clone(), false);
     let visitor = ResidencyVisitor {
         nr_entries: AtomicU32::new(0),
@@ -158,7 +157,7 @@ fn print_layout<V: Unpack>(bm: &Arc<BlockManager>, root: u64) {
     let _ = walker.walk::<LayoutVisitor, V>(&mut path, &visitor2, root);
 }
 
-fn residency<V: Unpack>(stats: &TreeStats) -> Result<usize> {
+pub fn residency<V: Unpack>(stats: &TreeStats) -> Result<usize> {
     let max_entries = calc_max_entries::<V>() as u64;
     let percent = (stats.nr_entries * 100) / (max_entries * stats.nr_leaves);
     Ok(percent as usize)
@@ -167,7 +166,7 @@ fn residency<V: Unpack>(stats: &TreeStats) -> Result<usize> {
 // Because this is a walk it implicitly checks the btree.  Returns
 // average residency as a _percentage_.
 pub fn calc_residency<V: Unpack>(bm: &Arc<BlockManager>, root: u64) -> Result<usize> {
-    let stats = get_stats::<V>(bm, root)?;
+    let stats = get_tree_stats::<V>(bm, root)?;
     debug!(
         "nr_leaves = {}, nr_entries = {}",
         stats.nr_leaves, stats.nr_entries
@@ -370,14 +369,14 @@ fn test_del_empty(fix: &mut Fixture) -> Result<()> {
     Ok(())
 }
 
-struct TreeStats {
-    nr_leaves: u64,
-    nr_entries: u64,
+pub struct TreeStats {
+    pub nr_leaves: u64,
+    pub nr_entries: u64,
 }
 
 #[allow(dead_code)]
-struct BTreeTest<'a> {
-    fix: &'a mut Fixture,
+pub struct BTreeTest<'a> {
+    pub fix: &'a mut Fixture,
     bm: Addr,
     tm: Addr,
     sm: Addr,
@@ -388,7 +387,7 @@ struct BTreeTest<'a> {
 }
 
 impl<'a> BTreeTest<'a> {
-    fn new(fix: &'a mut Fixture) -> Result<Self> {
+    pub fn new(fix: &'a mut Fixture) -> Result<Self> {
         let bm = dm_bm_create(fix, 1024)?;
         let (tm, sm) = dm_tm_create(fix, bm, 0)?;
 
@@ -424,7 +423,7 @@ impl<'a> BTreeTest<'a> {
         })
     }
 
-    fn begin(&mut self) -> Result<()> {
+    pub fn begin(&mut self) -> Result<()> {
         if self.sb.is_some() {
             return Err(anyhow!("transaction already begun"));
         }
@@ -433,21 +432,14 @@ impl<'a> BTreeTest<'a> {
         Ok(())
     }
 
-    fn insert(&mut self, key: u64) -> Result<()> {
+    pub fn insert(&mut self, key: u64) -> Result<()> {
         let ks = vec![key];
         let v = Value64(key_to_value(key));
         self.root = dm_btree_insert(self.fix, &self.info, self.root, &ks, &v)?;
         Ok(())
     }
 
-    fn insert_kv(&mut self, key: u64, value: u64) -> Result<()> {
-        let ks = vec![key];
-        let v = Value64(value);
-        self.root = dm_btree_insert(self.fix, &self.info, self.root, &ks, &v)?;
-        Ok(())
-    }
-
-    fn lookup(&mut self, key: u64) -> Result<()> {
+    pub fn lookup(&mut self, key: u64) -> Result<()> {
         let keys = vec![key];
         let v = dm_btree_lookup(self.fix, &self.info, self.root, &keys)?;
         ensure!(v == Value64(key_to_value(key)));
@@ -456,30 +448,30 @@ impl<'a> BTreeTest<'a> {
 
     // This uses Rust code, rather than doing look ups via the kernel
     // code.
-    fn check_keys_present(&self, keys: &[u64]) -> Result<()> {
+    pub fn check_keys_present(&self, keys: &[u64]) -> Result<()> {
         let bm = get_bm(self.fix, self.bm);
         check_keys_present(&bm, self.root, keys)
     }
 
-    fn commit(&mut self) -> Result<()> {
+    pub fn commit(&mut self) -> Result<()> {
         dm_tm_pre_commit(self.fix, self.tm)?;
         dm_tm_commit(self.fix, self.tm, self.sb.unwrap())?;
         self.sb = None;
         Ok(())
     }
 
-    fn stats_start(&mut self) {
+    pub fn stats_start(&mut self) {
         let bm = get_bm(self.fix, self.bm);
         self.baseline = Stats::collect_stats(self.fix, &bm);
     }
 
-    fn stats_delta(&mut self) -> Result<Stats> {
+    pub fn stats_delta(&mut self) -> Result<Stats> {
         let bm = get_bm(self.fix, self.bm);
         let delta = self.baseline.delta(self.fix, &bm);
         Ok(delta)
     }
 
-    fn stats_report(&self, desc: &str, count: u64) -> Result<()> {
+    pub fn stats_report(&self, desc: &str, count: u64) -> Result<()> {
         let bm = get_bm(self.fix, self.bm);
         let delta = self.baseline.delta(self.fix, &bm);
         info!(
@@ -493,17 +485,21 @@ impl<'a> BTreeTest<'a> {
         Ok(())
     }
 
-    fn get_stats(&self) -> Result<TreeStats> {
+    pub fn get_tree_stats(&self) -> Result<TreeStats> {
         let bm = get_bm(self.fix, self.bm);
-        get_stats::<Value64>(&bm, self.root)
+        get_tree_stats::<Value64>(&bm, self.root)
     }
 
-    fn residency(&self) -> Result<usize> {
+    pub fn get_bm(&self) -> Arc<BlockManager> {
+        get_bm(self.fix, self.bm)
+    }
+
+    pub fn residency(&self) -> Result<usize> {
         let bm = get_bm(self.fix, self.bm);
         calc_residency::<Value64>(&bm, self.root)
     }
 
-    fn print_layout(&self) -> Result<()> {
+    pub fn print_layout(&self) -> Result<()> {
         let bm = get_bm(self.fix, self.bm);
         print_layout::<Value64>(&bm, self.root);
         Ok(())
@@ -618,294 +614,6 @@ fn test_insert_runs(fix: &mut Fixture) -> Result<()> {
     }
 
     do_insert_test_(fix, &shuffled_keys, 2, 80)
-}
-
-//-------------------------------
-
-fn do_insert_bench_(fix: &mut Fixture, keys: &[u64], commit_interval: usize) -> Result<()> {
-    standard_globals(fix)?;
-    let mut btree = BTreeTest::new(fix)?;
-
-    let mut dblocks: Vec<u64> = (0..keys.len() as u64).into_iter().collect();
-    let mut rng = rand_chacha::ChaCha8Rng::seed_from_u64(1);
-    dblocks.shuffle(&mut rng);
-
-    let mut csv = File::create("./btree.csv")?;
-    writeln!(
-        csv,
-        "inserts, nr_leaves, nr_entries, residency, instructions, read_locks, write_locks"
-    )?;
-
-    let mut total = 0;
-    let mut i = 0;
-    for chunk in keys.chunks(commit_interval) {
-        btree.stats_start();
-
-        btree.begin()?;
-        for k in chunk {
-            let _nr_inserted = btree.insert_kv(*k, dblocks[i])?;
-        }
-        btree.commit()?;
-        i += 1;
-
-        let delta = btree.stats_delta()?;
-        let stats = btree.get_stats()?;
-
-        total += chunk.len();
-        writeln!(
-            csv,
-            "{}, {}, {}, {}, {}, {}, {}",
-            total,
-            stats.nr_leaves,
-            stats.nr_entries,
-            residency::<Value64>(&stats)?,
-            delta.instrs / chunk.len() as u64,
-            delta.read_locks / chunk.len() as u64,
-            delta.write_locks / chunk.len() as u64,
-        )?;
-    }
-
-    Ok(())
-}
-
-fn do_insert_perf_(fix: &mut Fixture, keys: &[u64], commit_interval: usize) -> Result<()> {
-    standard_globals(fix)?;
-    let mut btree = BTreeTest::new(fix)?;
-
-    let mut csv = File::create("./btree.csv")?;
-    writeln!(
-        csv,
-        "inserts, nr_leaves, nr_entries, residency, instructions, read_locks, write_locks"
-    )?;
-
-    let mut total = 0;
-    for chunk in keys.chunks(commit_interval) {
-        //for chunk in keys.chunks(commit_interval) {
-        btree.stats_start();
-
-        btree.begin()?;
-        for k in chunk {
-            println!("=== inserting {}, {} ===", k, total);
-            btree.stats_start();
-            let _nr_inserted = btree.insert(*k)?;
-            let delta = btree.stats_delta()?;
-            total += 1;
-
-            writeln!(
-                csv,
-                "{}, {}, {}, {}, {}, {}, {}",
-                total, 0, 0, 0, delta.instrs, delta.read_locks, delta.write_locks,
-            )?;
-        }
-        btree.commit()?;
-
-        let delta = btree.stats_delta()?;
-        let stats = btree.get_stats()?;
-
-        writeln!(
-            csv,
-            "{}, {}, {}, {}, {}, {}, {}",
-            total,
-            stats.nr_leaves,
-            stats.nr_entries,
-            residency::<Value64>(&stats)?,
-            delta.instrs / chunk.len() as u64,
-            delta.read_locks / chunk.len() as u64,
-            delta.write_locks / chunk.len() as u64,
-        )?;
-    }
-
-    Ok(())
-}
-
-// This function is used to trace down how the btree insertion cost is increased
-// as the btree becomes bigger.
-// - do commit at every insertion
-// - print the stats for the insert operation
-// - print the stats for the (insert + commit) operations
-// - print the btree layout at every insertion
-fn do_insert_perf_with_layout_(
-    fix: &mut Fixture,
-    keys: &[u64],
-    commit_interval: usize,
-) -> Result<()> {
-    standard_globals(fix)?;
-    let mut btree = BTreeTest::new(fix)?;
-
-    let mut csv = File::create("./btree.csv")?;
-    writeln!(
-        csv,
-        "inserts, nr_leaves, nr_entries, residency, instructions, read_locks, write_locks"
-    )?;
-
-    let mut total = 0;
-    for k in keys {
-        btree.stats_start();
-
-        btree.begin()?;
-        {
-            println!("=== inserting {}, {} ===", k, total);
-            btree.stats_start();
-            let _nr_inserted = btree.insert(*k)?;
-            let delta = btree.stats_delta()?;
-            total += 1;
-
-            writeln!(
-                csv,
-                "{}, {}, {}, {}, {}, {}, {}",
-                total, 0, 0, 0, delta.instrs, delta.read_locks, delta.write_locks,
-            )?;
-        }
-        btree.commit()?;
-
-        let delta = btree.stats_delta()?;
-        let stats = btree.get_stats()?;
-        btree.print_layout()?;
-
-        writeln!(
-            csv,
-            "{}, {}, {}, {}, {}, {}, {}",
-            total,
-            stats.nr_leaves,
-            stats.nr_entries,
-            residency::<Value64>(&stats)?,
-            delta.instrs,
-            delta.read_locks,
-            delta.write_locks,
-        )?;
-    }
-
-    Ok(())
-}
-
-// do the insertion benchmark, and print the btree layout at every commit as well
-fn do_insert_bench_with_layout_(
-    fix: &mut Fixture,
-    keys: &[u64],
-    commit_interval: usize,
-) -> Result<()> {
-    standard_globals(fix)?;
-    let mut btree = BTreeTest::new(fix)?;
-
-    let mut dblocks: Vec<u64> = (0..keys.len() as u64).into_iter().collect();
-    let mut rng = rand_chacha::ChaCha8Rng::seed_from_u64(1);
-    dblocks.shuffle(&mut rng);
-
-    let mut csv = File::create("./btree.csv")?;
-    writeln!(
-        csv,
-        "inserts, nr_leaves, nr_entries, residency, instructions, read_locks, write_locks"
-    )?;
-
-    let mut total = 0;
-    let mut i = 0;
-    for chunk in keys.chunks(commit_interval) {
-        btree.stats_start();
-
-        btree.begin()?;
-        for k in chunk {
-            let _nr_inserted = btree.insert_kv(*k, dblocks[i])?;
-        }
-        btree.commit()?;
-        i += 1;
-
-        println!("=== round {} ===", total);
-
-        let delta = btree.stats_delta()?;
-        let stats = btree.get_stats()?;
-
-        total += chunk.len();
-        writeln!(
-            csv,
-            "{}, {}, {}, {}, {}, {}, {}",
-            total,
-            stats.nr_leaves,
-            stats.nr_entries,
-            residency::<Value64>(&stats)?,
-            delta.instrs / chunk.len() as u64,
-            delta.read_locks / chunk.len() as u64,
-            delta.write_locks / chunk.len() as u64,
-        )?;
-
-        btree.print_layout();
-    }
-
-    Ok(())
-}
-
-fn do_lookup_bench_(fix: &mut Fixture, keys: &[u64], commit_interval: usize) -> Result<()> {
-    standard_globals(fix)?;
-    let mut btree = BTreeTest::new(fix)?;
-
-    let mut csv = File::create("./btree.csv")?;
-    writeln!(
-        csv,
-        "nr_inserted, nr_leaves, nr_entries, residency, instructions, read_locks, write_locks"
-    )?;
-
-    let mut total = 0;
-    for chunk in keys.chunks(commit_interval) {
-        btree.begin()?;
-        for k in chunk {
-            let _nr_inserted = btree.insert(*k)?;
-        }
-        btree.commit()?;
-
-        btree.stats_start();
-        for k in chunk {
-            let _ = btree.lookup(*k)?;
-        }
-        let delta = btree.stats_delta()?;
-        let stats = btree.get_stats()?;
-
-        total += chunk.len();
-        writeln!(
-            csv,
-            "{}, {}, {}, {}, {}, {}, {}",
-            total,
-            stats.nr_leaves,
-            stats.nr_entries,
-            residency::<Value64>(&stats)?,
-            delta.instrs / chunk.len() as u64,
-            delta.read_locks / chunk.len() as u64,
-            delta.write_locks / chunk.len() as u64,
-        )?;
-    }
-
-    Ok(())
-}
-
-fn bench_insert_random(fix: &mut Fixture) -> Result<()> {
-    let mut keys: Vec<u64> = (0..200000).rev().collect();
-    let mut rng = rand_chacha::ChaCha8Rng::seed_from_u64(1);
-    keys.shuffle(&mut rng);
-    do_insert_bench_(fix, &keys, 100)
-    //do_insert_perf_(fix, &keys, 100)
-}
-
-fn bench_insert_random_with_layout(fix: &mut Fixture) -> Result<()> {
-    let mut keys: Vec<u64> = (0..200000).rev().collect();
-    let mut rng = rand_chacha::ChaCha8Rng::seed_from_u64(1);
-    keys.shuffle(&mut rng);
-    //do_insert_bench_(fix, &keys, 100)
-    do_insert_bench_with_layout_(fix, &keys, 100)
-}
-
-fn bench_insert_ascending(fix: &mut Fixture) -> Result<()> {
-    let keys: Vec<u64> = (0..5000).collect();
-    do_insert_perf_(fix, &keys, 100)
-}
-
-fn bench_insert_ascending_with_layout(fix: &mut Fixture) -> Result<()> {
-    let keys: Vec<u64> = (0..5000).collect();
-    do_insert_perf_with_layout_(fix, &keys, 100)
-}
-
-fn bench_lookup_random(fix: &mut Fixture) -> Result<()> {
-    let mut keys: Vec<u64> = (0..200000).rev().collect();
-    let mut rng = rand_chacha::ChaCha8Rng::seed_from_u64(1);
-    keys.shuffle(&mut rng);
-    do_lookup_bench_(fix, &keys, 100)
 }
 
 //-------------------------------
@@ -1227,59 +935,6 @@ pub fn register_tests(runner: &mut TestRunner) -> Result<()> {
             test!("left-only", test_redistribute3_left_only)
             test!("right-below-target", test_redistribute3_right_below_target)
             test!("right-only", test_redistribute3_right_only)
-        }
-    };
-
-    Ok(())
-}
-
-pub fn register_bench(runner: &mut TestRunner) -> Result<()> {
-    let kmodules = vec![PDATA_MOD];
-    let mut prefix: Vec<&'static str> = Vec::new();
-
-    macro_rules! test_section {
-        ($path:expr, $($s:stmt)*) => {{
-            prefix.push($path);
-            $($s)*
-            prefix.pop().unwrap();
-        }}
-    }
-
-    macro_rules! test {
-        ($path:expr, $func:expr) => {{
-            prefix.push($path);
-            let p = prefix.concat();
-            prefix.pop().unwrap();
-            runner.register(&p, Test::new(kmodules.clone(), Box::new($func)));
-        }};
-    }
-
-    test_section! {
-        "/pdata/btree/",
-
-        test_section! {
-            "insert/",
-            test!("random", bench_insert_random)
-        }
-
-        test_section! {
-            "insert/",
-            test!("random_layout", bench_insert_random_with_layout)
-        }
-
-        test_section! {
-            "insert/",
-            test!("ascending", bench_insert_ascending)
-        }
-
-        /*test_section! {
-            "insert/",
-            test!("ascending_layout", bench_insert_ascending_with_layout)
-        }*/
-
-        test_section! {
-            "lookup/",
-            test!("random", bench_lookup_random)
         }
     };
